Disable TGM and rename admin menu items. Closes #104
