+++ assets/css/style.css
@@ -730,7 +730,3 @@ button.close-ticket:hover {
     font-size: 12px;
     color: #9d9d9d;
 }
-
-.footer-text a {
-    color: #9d9d9d;
-}
